# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Journal Management Admin API",
    "description": "期刊管理系统后台接口文档。包含用户、文章、期刊、通知、联系等模块。",
    "contact": {
      "name": "KSTT",
      "email": "support@kstt.com"
    },
    "license": {
      "name": "Apache 2.0"
    },
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:10086/",
      "description": "默认服务器"
    },
    {
      "url": "http://localhost:10086/",
      "description": "本地开发"
    }
  ],
...
  "paths": {
...
    "/api/sys/menu": {
      "put": {
        "tags": [
          "系统菜单管理"
        ],
        "summary": "修改系统菜单",
        "description": "更新菜单信息，需要提供完整的菜单信息，包括菜单ID",
        "operationId": "edit_6",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SysMenu"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "服务器内部错误",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "请求参数无效",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "404": {
            "description": "菜单不存在",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "修改成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "系统菜单管理"
        ],
        "summary": "新增系统菜单",
        "description": "创建新的系统菜单，需要提供菜单名称、类型、路由、权限标识等信息",
        "operationId": "add_6",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SysMenu"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "服务器内部错误",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "请求参数无效或菜单路由已存在",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "新增成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/articles/keywords/{keywordId}": {
      "get": {
        "tags": [
          "文章关键词管理"
        ],
        "summary": "根据关键词ID获取关键词信息",
        "description": "查询单个关键词的详细信息",
        "operationId": "getInfo",
        "parameters": [
          {
            "name": "keywordId",
            "in": "path",
            "description": "关键词ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/articles/files/{fileId}": {
      "get": {
        "tags": [
          "文章提交文件管理"
        ],
        "summary": "根据文件ID获取文件信息",
        "description": "查询单个文件的详细信息",
        "operationId": "getInfo_1",
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "description": "文件ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/team/{teamId}": {
      "get": {
        "tags": [
          "编辑团队管理"
        ],
        "summary": "获取编辑团队成员详细信息",
        "description": "根据团队成员ID获取详细信息，包括姓名、职位、简介等",
        "operationId": "getInfo_2",
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "description": "团队成员ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "服务器内部错误",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "获取成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AboutEditorialTeam"
                }
              }
            }
          },
          "404": {
            "description": "团队成员不存在",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sys/user/{userId}": {
      "get": {
        "tags": [
          "系统用户管理"
        ],
        "summary": "获取系统用户详细信息",
        "description": "根据用户ID获取完整的用户信息，包括基本信息、角色信息等（密码字段不会返回）",
        "operationId": "getInfo_3",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "用户ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "服务器内部错误",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "获取成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SysUser"
                }
              }
            }
          },
          "404": {
            "description": "用户不存在",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sys/subject/{subjectId}": {
      "get": {
        "tags": [
          "学科管理"
        ],
        "summary": "获取学科详细信息",
        "description": "根据学科ID获取学科详细信息",
        "operationId": "getInfo_4",
        "parameters": [
          {
            "name": "subjectId",
            "in": "path",
            "description": "学科ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "学科管理"
        ],
        "summary": "删除学科",
        "description": "根据学科ID删除学科信息",
        "operationId": "remove_4",
        "parameters": [
          {
            "name": "subjectId",
            "in": "path",
            "description": "学科ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sys/role/{roleId}": {
      "get": {
        "tags": [
          "系统角色管理"
        ],
        "summary": "获取系统角色详细信息",
        "description": "根据角色ID获取角色的详细信息",
        "operationId": "getInfo_5",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "description": "角色ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "服务器内部错误",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "获取成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SysRole"
                }
              }
            }
          },
          "404": {
            "description": "角色不存在",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sys/menu/{menuId}": {
      "get": {
        "tags": [
          "系统菜单管理"
        ],
        "summary": "获取系统菜单详细信息",
        "description": "根据菜单ID获取菜单的详细信息，包括父级菜单、权限标识等",
        "operationId": "getInfo_6",
        "parameters": [
          {
            "name": "menuId",
            "in": "path",
            "description": "菜单ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "服务器内部错误",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "404": {
            "description": "菜单不存在",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "获取成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SysMenu"
                }
              }
            }
          }
        }
      }
    },
    "/api/sys/menu/list": {
      "get": {
        "tags": [
          "系统菜单管理"
        ],
        "summary": "查询系统菜单列表",
        "description": "分页查询系统菜单列表，支持按菜单名称、类型、状态等条件筛选",
        "operationId": "list_6",
        "parameters": [
          {
            "name": "menu",
            "in": "query",
            "description": "查询条件（可按菜单名称、类型等筛选）",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SysMenu"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "服务器内部错误",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "查询成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TableDataInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/sys/info/{infoId}": {
      "get": {
        "tags": [
          "系统信息管理"
        ],
        "summary": "获取系统信息详细信息",
        "description": "根据ID获取期刊系统的详细信息，包括ISSN、影响因子、官网等",
        "operationId": "getInfo_7",
        "parameters": [
          {
            "name": "infoId",
            "in": "path",
            "description": "系统信息ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "服务器内部错误",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "获取成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SysInfo"
                }
              }
            }
          },
          "404": {
            "description": "信息不存在",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sys/area/{areaId}": {
      "get": {
        "tags": [
          "学科领域管理"
        ],
        "summary": "获取领域详细信息",
        "description": "根据领域ID获取领域详细信息",
        "operationId": "getInfo_8",
        "parameters": [
          {
            "name": "areaId",
            "in": "path",
            "description": "领域ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "学科领域管理"
        ],
        "summary": "删除领域",
        "description": "根据领域ID删除学科领域信息",
        "operationId": "remove_8",
        "parameters": [
          {
            "name": "areaId",
            "in": "path",
            "description": "领域ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sys/academic-degree/{degreeId}": {
      "get": {
        "tags": [
          "学术学位管理"
        ],
        "summary": "获取学位详细信息",
        "description": "根据学位ID获取学位详细信息",
        "operationId": "getInfo_9",
        "parameters": [
          {
            "name": "degreeId",
            "in": "path",
            "description": "学位ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notices/{noticeId}": {
      "get": {
        "tags": [
          "通知管理"
        ],
        "summary": "获取通知详细信息",
        "description": "根据通知ID获取通知的详细内容",
        "operationId": "getInfo_10",
        "parameters": [
          {
            "name": "noticeId",
            "in": "path",
            "description": "通知ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "服务器内部错误",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "获取成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Notice"
                }
              }
            }
          },
          "404": {
            "description": "通知不存在",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/journal/{journalId}": {
      "get": {
        "tags": [
          "期刊管理"
        ],
        "summary": "获取期刊详细信息",
        "description": "根据期刊ID获取完整的期刊信息，包括卷次、期号、出版日期、主题、简介、封面、PDF文件等详细信息以及关联的文章数量",
        "operationId": "getInfo_11",
        "parameters": [
          {
            "name": "journalId",
            "in": "path",
            "description": "期刊ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "服务器内部错误",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "获取成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/JournalDetailDTO"
                }
              }
            }
          },
          "404": {
            "description": "期刊不存在",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/contact/message/{messageId}": {
      "get": {
        "tags": [
          "联系消息管理"
        ],
        "summary": "获取联系消息详细信息",
        "description": "根据消息ID获取消息的详细内容，包括发送者信息、消息内容、回复等",
        "operationId": "getInfo_12",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "description": "消息ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "服务器内部错误",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "404": {
            "description": "消息不存在",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "获取成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ContactMessage"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/contact/faq/{faqId}": {
      "get": {
        "tags": [
          "常见问题管理"
        ],
        "summary": "获取FAQ详细信息",
        "description": "根据FAQ ID获取问题的详细内容和答案",
        "operationId": "getInfo_13",
        "parameters": [
          {
            "name": "faqId",
            "in": "path",
            "description": "FAQ ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "服务器内部错误",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "获取成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ContactFaq"
                }
              }
            }
          },
          "404": {
            "description": "FAQ不存在",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/articles/{articleId}": {
      "get": {
        "tags": [
          "文章管理"
        ],
        "summary": "根据文章ID获取文章详情",
        "description": "根据文章ID获取完整的文章信息，包括标题、摘要、关键词、状态、作者、关联期刊等详细信息",
        "operationId": "getInfo_14",
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "description": "文章ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "服务器内部错误",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          },
          "200": {
            "description": "获取成功",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "404": {
            "description": "文章不存在",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AjaxResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SysMenu": {
        "type": "object",
        "properties": {
          "createBy": {
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateBy": {
            "type": "string"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "remark": {
            "type": "string"
          },
          "menuId": {
            "type": "integer",
            "format": "int64"
          },
          "menuName": {
            "type": "string"
          },
          "menuParentId": {
            "type": "integer",
            "format": "int64"
          },
          "menuOrderNum": {
            "type": "integer",
            "format": "int32"
          },
          "menuPath": {
            "type": "string"
          },
          "menuComponent": {
            "type": "string"
          },
          "menuType": {
            "type": "string"
          },
          "menuVisible": {
            "type": "integer",
            "format": "int32"
          },
          "menuStatus": {
            "type": "integer",
            "format": "int32"
          },
          "menuPerms": {
            "type": "string"
          },
          "menuIcon": {
            "type": "string"
          }
        }
      },
...
    }
  }
}